Move collect_simple_response helper to core.testing
