+++ agents-core/vision_agents/testing/__init__.py
@@ -42,6 +42,7 @@ from vision_agents.testing._events import (
 from vision_agents.testing._judge import Judge, JudgeVerdict, LLMJudge
 from vision_agents.testing._run_result import TestResponse
 from vision_agents.testing._session import TestSession
+from vision_agents.testing._utils import collect_simple_response
 
 __all__ = [
     "Judge",
@@ -53,4 +54,5 @@ __all__ = [
     "FunctionCallEvent",
     "FunctionCallOutputEvent",
     "RunEvent",
+    "collect_simple_response",
 ]
