+++ plugins/aws/tests/test_aws.py
@@ -7,7 +7,7 @@ from dotenv import load_dotenv
 from vision_agents.core.agents.conversation import InMemoryConversation
 from vision_agents.plugins.aws.aws_llm import BedrockLLM
 
-from tests.utils import collect_simple_response
+from vision_agents.testing import collect_simple_response
 
 load_dotenv()
