+++ plugins/gemini/tests/test_gemini_vlm.py
@@ -7,7 +7,7 @@ from dotenv import load_dotenv
 from vision_agents.core.agents.conversation import InMemoryConversation
 from vision_agents.plugins.gemini import VLM
 
-from tests.utils import collect_simple_response
+from vision_agents.testing import collect_simple_response
 
 load_dotenv()
