+++ plugins/huggingface/tests/test_transformers_llm.py
@@ -7,7 +7,7 @@ import pytest
 import torch
 from conftest import skip_blockbuster
 from vision_agents.core.agents.conversation import InMemoryConversation
-from tests.utils import collect_simple_response
+from vision_agents.testing import collect_simple_response
 from vision_agents.core.llm.llm import LLMResponseFinal
 from vision_agents.plugins.huggingface.transformers_llm import (
     ModelResources,
