+++ plugins/nvidia/tests/test_nvidia_vlm.py
@@ -19,7 +19,7 @@ from vision_agents.core.agents.conversation import InMemoryConversation
 from vision_agents.core.edge.types import Participant
 from vision_agents.plugins.nvidia import VLM
 
-from tests.utils import collect_simple_response
+from vision_agents.testing import collect_simple_response
 
 load_dotenv()
